refactor: move createCars from CarService to CarFactory
