+++ README.md
@@ -50,6 +50,7 @@
 - **Controller**: `RacingCarController`클래스는 전체 게임의 흐름을 제어합니다.
     - `InputView`를 통해 자동차 이름과 이동 횟수 입력을 받습니다.
     - `InputValidator`를 통해 입력값의 유효성 검사를 합니다.
-    - `CarService `를 통해 자동차 객체를 생성하고, 레이싱 게임을 진행하고, 게임 결과를 출력합니다.
+    - `CarFactory`를 통해 자동차 객체를 생성한다.
+    - `CarService `를 통해 레이싱 게임을 진행하고, 게임 결과를 출력합니다.
 - **Model**: `Car`클래스는 자동차의 상태(이름, 현재위치)를 표현하고 동작(전진)을 정의합니다.
 - **View**: `InputView`클래스는 사용자로부터 입력값을 받습니다.
