+++ src/main/java/racingcar/service/CarFactory.java
@@ -1,0 +1,16 @@
+package racingcar.service;
+
+import racingcar.model.Car;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CarFactory {
+    public static List<Car> createCars(String[] carNames) {
+        List<Car> cars = new ArrayList<>();
+        for (String name : carNames) {
+            cars.add(new Car(name));
+        }
+        return cars;
+    }
+}
